pyproject.toml
[project]
name = "ai-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "openai>=2.15.0",
    "pytest>=9.0.2",
    "python-dotenv==1.1.0",
]
